+++ kornia/geometry/transform/__init__.py
@@ -6,6 +6,5 @@ from .flips import *
 from .homography_warper import *
 from .image_registrator import *
 from .imgwarp import *
-from .projwarp import *
 from .pyramid import *
 from .thin_plate_spline import *
